requirements.txt
ipython
pytest-selenium==1.2.1
pytest-xdist==1.14
